pyproject.toml
[tool.poetry]
name = "bigquery-to-datastore-via-beam"
version = "0.1.0"
description = ""
authors = ["Richard He"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apache-beam = { version = "2.51.0", extras = ["gcp"] }
google-cloud-bigquery = "3.13.0"
google-cloud-bigquery-storage = "2.22.0"
PyYAML = "6.0.1"
faker = "^20.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
